Combine common BaseDB code.
